Broke the integration tests out into individual sections
